Enhancements to build:
* Reorganized docker tools into docker subparser to improve clarity, and
renamed `hydrant publish` -> `hydrant docker push`
* build is now aware of context, can be used from both workflow and task
directories, and can build multiple images when used from the workflow
directory. The help message for build will vary depending if user is
inside a workflow dir, a task dir, or neither, telling the user what
task(s) is/are available to build.
* init now prepends the workflow name to the default numbered task names
so that users may choose to not rename tasks if they so desire.